refactor(runtime): split class.rs into cohesive modules (§7-8)
`class.rs` was 1182 lines, over the 500-line module limit. Split it into three
files along its natural seams, each under 460 lines, with no behavior change:

- `class.rs` (300) — class lifecycle and structure: instance DESTROY running
  (`run_pending_instance_destroys*`), role-method composition-conflict detection,
  C3 MRO computation, and `format_method_candidate_signatures` for dispatch error
  messages.
- `class_introspection.rs` (446) — native-method tables (`is_native_method`),
  user-method/accessor presence checks, class-level attribute lookup/mutation,
  and attribute collection across the MRO and composed roles.
- `class_dispatch.rs` (457) — instance-method dispatch: candidate resolution,
  `nextsame`/`samewith` frame + method-wrap chain setup (`run_instance_method`),
  compiled-bytecode candidate execution
  (`run_resolved_method_compiled_or_treewalk`), and `handles`-delegation
  forwarding (`forward_resolved_delegation`).

The only private fn (`run_pending_instance_destroys_inner`) stays in `class.rs`
with its sole caller. The `use crate::symbol::Symbol` import moves to
`class_dispatch` (its only user). Third slice of the §7-8 file-splitting sweep.
